invite_users/requirements.txt
asgiref==3.5.2
gunicorn==20.0.4
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.1
django-utils-six==2.0
pycodestyle==2.9.1
python-dotenv==0.20.0
sqlparse==0.4.2
toml==0.10.2
psycopg2-binary==2.8.6
